+++ lib/events/events.dart
@@ -1,13 +1,17 @@
 // UI Events
 import 'package:event_bloc/event_bloc.dart';
 import 'package:flutter/material.dart';
+import 'package:vhcsite/bloc/blog.dart';
 
 enum VHCSiteEvent<T> {
   button<String>(),
   setTextScale<double>(),
+
   loadBlog<void>(),
   pickBlogCategory<String?>(),
   changeBlogSearchTerm<String?>(),
+  pickBlogSortOrder<BlogSortOrder>(),
+  clearCategoryFilters<void>(),
   ;
 
   BlocEventType<T> get event => BlocEventType.fromObject(this);